idcard: protect against stray files in the image dir
OSX seems to drop extra files in directories that os.listdir()
seems to want to include. So the assumption that our image directory
only contains images is false.

Instead, glob for files that look like the image files we are expecting
to find in there so we can filter out some of the rubbish.

Signed-off-by: Neil Barber <nbarber@rtype.local>