idcard: understand eayso data that contains commas
The csv file is comma separated but we cannot split the csv on
commas because, it turns out, the data itself can contain commas.
This means splitting on commas can leave the data in an unstructured
state.

Split on "," instead. This keeps data elements that contain a comma
safely intact.

Signed-off-by: Neil Barber <nbarber@rtype.local>